Re-org map component
