+++ src/lib/Maps/SVGMap/LabelLayer.svelte
