+++ src/lib/Maps/SVGMap/PointLayer.docs.md
