+++ src/lib/Maps/SVGMap/PointLayer.svelte
