+++ src/lib/Maps/SVGMap/ZoomControls.svelte
