+++ src/lib/Maps/index.js
@@ -1,0 +1,4 @@
+export { default as SVGMap } from "./SVGMap/SVGMap.svelte";
+export { default as PolygonLayer } from "./SVGMap/PolygonLayer.svelte";
+export { default as PointLayer } from "./SVGMap/PointLayer.svelte";
+export { default as LabelLayer } from "./SVGMap/LabelLayer.svelte";
